+++ jest.config.js
@@ -1,3 +1,3 @@
-const { setupTestFrameworkScriptFile, ...config } = require('@socifi/jest-config')();
-
-module.exports = { ...config, transform: { '.*': 'babel-jest' } };
+module.exports = {
+    ...require('@socifi/jest-config')(),
+};