+++ src/types/api/Workouts.ts
@@ -2,5 +2,5 @@ import { Workout } from './Workout';
 import { Paging } from './Paging';
 
 export type Workouts = {
-    data: Array<Workout>,
+    data: Workout[],
 } & Paging;